fix(LottoServiceImpl, ApplicationTest): 수익률 계산 로직 오버플로우 fix
- int 범위를 넘어가는 경우를 대비하기 위해 당첨금 합산 로직에서 long을 사용하도록 변경

+++ src/main/java/lotto/back/lotto/service/LottoServiceImpl.java
@@ -79,8 +79,8 @@ public class LottoServiceImpl implements LottoService {
             throw new CustomIllegalArgumentException("구매 금액은 0보다 커야 합니다.");
         }
 
-        int sum = results.getLottoResults().stream()
-                .mapToInt(lottoResult -> lottoResult.getLottoStatus().getPrize())
+        long sum = results.getLottoResults().stream()
+                .mapToLong(lottoResult -> lottoResult.getLottoStatus().getPrize())
                 .sum();
 
         return (double) sum / price * 100;
